Define global variables with window instead of global scope.
- Please view the diff with Github's setting: "Hide whitespace changes"

Using the global variables defined in playcanvas-anim.js was impossible when trying to pack the scripts with something like webpack, since the file would no longer be executed in the global scope.
Ideally the variables should be in some better-named namespace (like gltf.AnimationClip or pc.gltf.AnimationClip), but putting them in window is a non-breaking change and at least makes them more accessible.
playcanvas-gltf.js did not have this problem, but the code was updated to match the style of playcanvas-anim.js.